add tick rate to gui

+++ Erbium/Erbium/Private/Misc.cpp
@@ -27,7 +27,7 @@ void* Misc::SendRequestNow(void* Arg1, void* MCPData, int)
 float Misc::GetMaxTickRate(UEngine* Engine, float DeltaTime, bool bAllowFrameRateSmoothing)
 {
 	// improper, DS is supposed to do hitching differently
-	return FConfiguration::MaxTickRate;
+	return (float)FConfiguration::MaxTickRate;
 	//return std::clamp(1.f / DeltaTime, 1.f, FConfiguration::MaxTickRate);
 }
